+++ components/ProjectsTable.tsx
@@ -1,5 +1,5 @@
 import classNames from '@/lib/classNames';
-import { project } from '@prisma/client';
+import { Project } from '@prisma/client';
 import Link from 'next/link';
 import OccurrenceCounterLabel from './CounterLabel';
 
@@ -9,12 +9,12 @@ const statusClasses = {
   red: 'text-rose-400 bg-rose-400/50',
 };
 
-export default function ProjectsTable({ projects }: { projects: project[] }) {
+export default function ProjectsTable({ projects }: { projects: Project[] }) {
   return (
     <ul role="list" className="divide-y divide-white/5">
       {projects.map((project) => (
         <li
-          key={project.id.toString()}
+          key={project.id}
           className="relative flex items-center space-x-4 px-4 py-4 transition-colors duration-200 hover:bg-airbroke-800 sm:px-6 lg:px-8"
         >
           <div className="min-w-0 flex-auto">
@@ -36,14 +36,14 @@ export default function ProjectsTable({ projects }: { projects: project[] }) {
               </span>
 
               <h2 className="relative min-w-0 text-sm font-semibold leading-6 text-white">
-                <Link href={`/projects/${project.id}/notices`} className="flex gap-x-2">
+                <Link href={`/projects/${project.id}`} className="flex gap-x-2">
                   <span className="truncate">{project.organization.toLowerCase()}</span>
                   <span className="text-gray-400">/</span>
                   <span className="cursor-pointer whitespace-nowrap">{project.name}</span>
                   <span className="absolute inset-0" />
                 </Link>
                 <Link
-                  href={`/projects/${project.id}/notices`}
+                  href={`/projects/${project.id}`}
                   className={classNames(
                     'absolute inset-0 flex gap-x-2 bg-gradient-to-r from-gray-400 bg-clip-text text-transparent opacity-0 transition-opacity duration-300 ease-in-out hover:opacity-100',
                     project.notices_count === BigInt(0) ? 'to-green-600' : 'to-red-600'
@@ -57,7 +57,7 @@ export default function ProjectsTable({ projects }: { projects: project[] }) {
             </div>
           </div>
 
-          <Link href={`/projects/${project.id}/notices`}>
+          <Link href={`/projects/${project.id}`}>
             <OccurrenceCounterLabel counter={project.notices_count} />
           </Link>
         </li>