+++ lib/actions/airbrakeActions.ts
@@ -4,7 +4,7 @@ import { prisma } from '@/lib/db';
 import { Notifier as AirbrakeNodeNotifier } from '@airbrake/node';
 import { revalidatePath } from 'next/cache';
 
-export async function sendAirbrakeNodeException(projectId: bigint, host: string) {
+export async function sendAirbrakeNodeException(projectId: string, host: string) {
   const project = await prisma.project.findUnique({ where: { id: projectId } });
   if (!project) {
     throw new Error('Project not found');
@@ -28,7 +28,7 @@ export async function sendAirbrakeNodeException(projectId: bigint, host: string)
 
   try {
     revalidatePath(`/projects`)
-    revalidatePath(`/projects/${project.id}/notices`)
+    revalidatePath(`/projects/${project.id}`)
   } catch (err) {
     console.warn(err);
   }