+++ lib/auth.ts
@@ -1,3 +1,4 @@
+import { PrismaAdapter } from "@next-auth/prisma-adapter";
 import type { NextAuthOptions, Profile } from "next-auth";
 import AtlassianProvider from "next-auth/providers/atlassian";
 import AzureADProvider from "next-auth/providers/azure-ad";
@@ -7,6 +8,7 @@ import GitlabProvider from "next-auth/providers/gitlab";
 import GoogleProvider from "next-auth/providers/google";
 import KeycloakProvider from "next-auth/providers/keycloak";
 import { Octokit } from "octokit";
+import { prisma } from "./db";
 
 type ExtendedProfile = Profile & { [key: string]: any; };
 
@@ -78,7 +80,22 @@ export const authOptions: NextAuthOptions = {
   },
   debug: process.env.NEXTAUTH_DEBUG === "true",
   providers: getProviders(),
+  adapter: PrismaAdapter(prisma),
+
   callbacks: {
+    async jwt({ token, account, user }) {
+      // Persist the user id to the token right after signin
+      if (user?.id) {
+        token.id = user.id
+      }
+      return token
+    },
+    async session({ session, token }) {
+      if (session.user) {
+        session.user.id = token.id as string
+      }
+      return session;
+    },
     async signIn({ account, profile }) {
       const extendedProfile = profile as ExtendedProfile;