+++ middleware.ts
@@ -2,15 +2,6 @@ export { default } from "next-auth/middleware";
 
 export const config = {
   matcher: [
-    /*
-     * Match all request paths except for the ones starting with:
-     * - api (API routes)
-     * - _next/static (static files)
-     * - _next/image (image optimization files)
-     * - favicon.ico (favicon file)
-     * - robots.txt (robots file)
-     * - the homepage ("/")
-     */
-    '/((?!api|_next/|_static|icon|robots.txt|sitemap.xml|graphql).+)',
+    '/((?!api|_next/|notifier_api|icon|robots.txt|sitemap.xml|graphql).+)',
   ],
 };